@@ -1,20 +1,16 @@
-var React = require('react-native');
-var {
+import React from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
-  Text,
   Dimensions,
-  TouchableOpacity,
   Animated,
-  Platform,
-} = React;
+} from 'react-native';
 
-var MapView = require('react-native-maps');
-var PanController = require('./PanController');
-var PriceMarker = require('./AnimatedPriceMarker');
+import MapView from 'react-native-maps';
+import PanController from './PanController';
+import PriceMarker from './AnimatedPriceMarker';
 
-var screen = Dimensions.get('window');
+const screen = Dimensions.get('window');
 
 const ASPECT_RATIO = screen.width / screen.height;
 const LATITUDE = 37.78825;
@@ -22,19 +18,113 @@ const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 
-var ITEM_SPACING = 10;
-var ITEM_PREVIEW = 10;
-var ITEM_WIDTH = screen.width - 2 * ITEM_SPACING - 2 * ITEM_PREVIEW;
-var SNAP_WIDTH = ITEM_WIDTH + ITEM_SPACING;
-var ITEM_PREVIEW_HEIGHT = 150;
-var SCALE_END = screen.width / ITEM_WIDTH;
-var BREAKPOINT1 = 246;
-var BREAKPOINT2 = 350;
+const ITEM_SPACING = 10;
+const ITEM_PREVIEW = 10;
+const ITEM_WIDTH = screen.width - (2 * ITEM_SPACING) - (2 * ITEM_PREVIEW);
+const SNAP_WIDTH = ITEM_WIDTH + ITEM_SPACING;
+const ITEM_PREVIEW_HEIGHT = 150;
+const SCALE_END = screen.width / ITEM_WIDTH;
+const BREAKPOINT1 = 246;
+const BREAKPOINT2 = 350;
+const ONE = new Animated.Value(1);
 
-const ANDROID = Platform.OS === 'android';
+function getMarkerState(panX, panY, scrollY, i) {
+  const xLeft = (-SNAP_WIDTH * i) + (SNAP_WIDTH / 2);
+  const xRight = (-SNAP_WIDTH * i) - (SNAP_WIDTH / 2);
+  const xPos = -SNAP_WIDTH * i;
 
-var AnimatedViews = React.createClass({
-  getInitialState() {
+  const isIndex = panX.interpolate({
+    inputRange: [xRight - 1, xRight, xLeft, xLeft + 1],
+    outputRange: [0, 1, 1, 0],
+    extrapolate: 'clamp',
+  });
+
+  const isNotIndex = panX.interpolate({
+    inputRange: [xRight - 1, xRight, xLeft, xLeft + 1],
+    outputRange: [1, 0, 0, 1],
+    extrapolate: 'clamp',
+  });
+
+  const center = panX.interpolate({
+    inputRange: [xPos - 10, xPos, xPos + 10],
+    outputRange: [0, 1, 0],
+    extrapolate: 'clamp',
+  });
+
+  const selected = panX.interpolate({
+    inputRange: [xRight, xPos, xLeft],
+    outputRange: [0, 1, 0],
+    extrapolate: 'clamp',
+  });
+
+  const translateY = Animated.multiply(isIndex, panY);
+
+  const translateX = panX;
+
+  const anim = Animated.multiply(isIndex, scrollY.interpolate({
+    inputRange: [0, BREAKPOINT1],
+    outputRange: [0, 1],
+    extrapolate: 'clamp',
+  }));
+
+  const scale = Animated.add(ONE, Animated.multiply(isIndex, scrollY.interpolate({
+    inputRange: [BREAKPOINT1, BREAKPOINT2],
+    outputRange: [0, SCALE_END - 1],
+    extrapolate: 'clamp',
+  })));
+
+  // [0 => 1]
+  let opacity = scrollY.interpolate({
+    inputRange: [BREAKPOINT1, BREAKPOINT2],
+    outputRange: [0, 1],
+    extrapolate: 'clamp',
+  });
+
+  // if i === index: [0 => 0]
+  // if i !== index: [0 => 1]
+  opacity = Animated.multiply(isNotIndex, opacity);
+
+
+  // if i === index: [1 => 1]
+  // if i !== index: [1 => 0]
+  opacity = opacity.interpolate({
+    inputRange: [0, 1],
+    outputRange: [1, 0],
+  });
+
+  let markerOpacity = scrollY.interpolate({
+    inputRange: [0, BREAKPOINT1],
+    outputRange: [0, 1],
+    extrapolate: 'clamp',
+  });
+
+  markerOpacity = Animated.multiply(isNotIndex, markerOpacity).interpolate({
+    inputRange: [0, 1],
+    outputRange: [1, 0],
+  });
+
+  const markerScale = selected.interpolate({
+    inputRange: [0, 1],
+    outputRange: [1, 1.2],
+  });
+
+  return {
+    translateY,
+    translateX,
+    scale,
+    opacity,
+    anim,
+    center,
+    selected,
+    markerOpacity,
+    markerScale,
+  };
+}
+
+class AnimatedViews extends React.Component {
+  constructor(props) {
+    super(props);
+
     const panX = new Animated.Value(0);
     const panY = new Animated.Value(0);
 
@@ -59,8 +149,6 @@ var AnimatedViews = React.createClass({
       outputRange: [0, -100],
       extrapolate: 'clamp',
     });
-
-    const ONE = new Animated.Value(1);
 
     const markers = [
       {
@@ -89,100 +177,10 @@ var AnimatedViews = React.createClass({
       },
     ];
 
-    const animations = markers.map((m, i)  => {
-      const xLeft = -SNAP_WIDTH * i + SNAP_WIDTH/2;
-      const xRight = -SNAP_WIDTH * i - SNAP_WIDTH/2;
-      const xPos = -SNAP_WIDTH * i;
+    const animations = markers.map((m, i) =>
+      getMarkerState(panX, panY, scrollY, i));
 
-      const isIndex = panX.interpolate({
-        inputRange: [xRight - 1, xRight, xLeft, xLeft+1],
-        outputRange: [0, 1, 1, 0],
-        extrapolate: 'clamp',
-      });
-
-      const isNotIndex = panX.interpolate({
-        inputRange: [xRight - 1, xRight, xLeft, xLeft+1],
-        outputRange: [1, 0, 0, 1],
-        extrapolate: 'clamp',
-      });
-
-      const center = panX.interpolate({
-        inputRange: [xPos - 10, xPos, xPos + 10],
-        outputRange: [0, 1, 0],
-        extrapolate: 'clamp',
-      });
-
-      const selected = panX.interpolate({
-        inputRange: [xRight, xPos, xLeft],
-        outputRange: [0, 1, 0],
-        extrapolate: 'clamp',
-      });
-
-      const translateY = Animated.multiply(isIndex, panY);
-
-      const translateX = panX;
-
-      const anim = Animated.multiply(isIndex, scrollY.interpolate({
-        inputRange: [0, BREAKPOINT1],
-        outputRange: [0, 1],
-        extrapolate: 'clamp',
-      }));
-
-      const scale = Animated.add(ONE, Animated.multiply(isIndex, scrollY.interpolate({
-        inputRange: [BREAKPOINT1, BREAKPOINT2],
-        outputRange: [0, SCALE_END-1],
-        extrapolate: 'clamp',
-      })));
-
-      // [0 => 1]
-      var opacity = scrollY.interpolate({
-        inputRange: [BREAKPOINT1, BREAKPOINT2],
-        outputRange: [0, 1],
-        extrapolate: 'clamp',
-      });
-
-      // if i === index: [0 => 0]
-      // if i !== index: [0 => 1]
-      opacity = Animated.multiply(isNotIndex, opacity);
-
-
-      // if i === index: [1 => 1]
-      // if i !== index: [1 => 0]
-      opacity = opacity.interpolate({
-        inputRange: [0, 1],
-        outputRange: [1, 0],
-      });
-
-      var markerOpacity = scrollY.interpolate({
-        inputRange: [0, BREAKPOINT1],
-        outputRange: [0, 1],
-        extrapolate: 'clamp',
-      });
-
-      markerOpacity = Animated.multiply(isNotIndex, markerOpacity).interpolate({
-        inputRange: [0, 1],
-        outputRange: [1, 0],
-      });
-
-      var markerScale = selected.interpolate({
-        inputRange: [0, 1],
-        outputRange: [1, 1.2],
-      });
-
-      return {
-        translateY,
-        translateX,
-        scale,
-        opacity,
-        anim,
-        center,
-        selected,
-        markerOpacity,
-        markerScale,
-      };
-    });
-
-    return {
+    this.state = {
       panX,
       panY,
       animations,
@@ -193,17 +191,17 @@ var AnimatedViews = React.createClass({
       scale,
       translateY,
       markers,
-      region: new Animated.Region({
+      region: new MapView.AnimatedRegion({
         latitude: LATITUDE,
         longitude: LONGITUDE,
         latitudeDelta: LATITUDE_DELTA,
         longitudeDelta: LONGITUDE_DELTA,
       }),
     };
-  },
+  }
 
   componentDidMount() {
-    var { region, panX, panY, scrollX, markers } = this.state;
+    const { region, panX, panY, scrollX, markers } = this.state;
 
     panX.addListener(this.onPanXChange);
     panY.addListener(this.onPanYChange);
@@ -220,49 +218,52 @@ var AnimatedViews = React.createClass({
       }),
       duration: 0,
     }).start();
-  },
+  }
 
-  onStartShouldSetPanResponder(e) {
+  onStartShouldSetPanResponder = (e) => {
     // we only want to move the view if they are starting the gesture on top
     // of the view, so this calculates that and returns true if so. If we return
     // false, the gesture should get passed to the map view appropriately.
-    var { panY } = this.state;
-    var { pageY } = e.nativeEvent;
-    var topOfMainWindow = ITEM_PREVIEW_HEIGHT + 1 * panY.__getValue();
-    var topOfTap = screen.height - pageY;
+    const { panY } = this.state;
+    const { pageY } = e.nativeEvent;
+    const topOfMainWindow = ITEM_PREVIEW_HEIGHT + panY.__getValue();
+    const topOfTap = screen.height - pageY;
 
     return topOfTap < topOfMainWindow;
-  },
+  }
 
-  onMoveShouldSetPanResponder(e) {
-    var { panY } = this.state;
-    var { pageY } = e.nativeEvent;
-    var topOfMainWindow = ITEM_PREVIEW_HEIGHT + 1 * panY.__getValue();
-    var topOfTap = screen.height - pageY;
+  onMoveShouldSetPanResponder = (e) => {
+    const { panY } = this.state;
+    const { pageY } = e.nativeEvent;
+    const topOfMainWindow = ITEM_PREVIEW_HEIGHT + panY.__getValue();
+    const topOfTap = screen.height - pageY;
 
     return topOfTap < topOfMainWindow;
-  },
+  }
 
-  onPanXChange({ value }) {
-    var { index, region, panX, markers } = this.state;
-    var newIndex = Math.floor((-1 * value + SNAP_WIDTH / 2) / SNAP_WIDTH);
+  onPanXChange = ({ value }) => {
+    const { index } = this.state;
+    const newIndex = Math.floor(((-1 * value) + (SNAP_WIDTH / 2)) / SNAP_WIDTH);
     if (index !== newIndex) {
       this.setState({ index: newIndex });
     }
-  },
+  }
 
-  onPanYChange({ value }) {
-    var { canMoveHorizontal, region, scrollY, scrollX, markers, index } = this.state;
-    var shouldBeMovable = Math.abs(value) < 2;
+  onPanYChange = ({ value }) => {
+    const { canMoveHorizontal, region, scrollY, scrollX, markers, index } = this.state;
+    const shouldBeMovable = Math.abs(value) < 2;
     if (shouldBeMovable !== canMoveHorizontal) {
       this.setState({ canMoveHorizontal: shouldBeMovable });
       if (!shouldBeMovable) {
-        var { coordinate } = markers[index];
+        const { coordinate } = markers[index];
         region.stopAnimation();
         region.timing({
           latitude: scrollY.interpolate({
             inputRange: [0, BREAKPOINT1],
-            outputRange: [coordinate.latitude, coordinate.latitude - LATITUDE_DELTA * 0.5 * 0.375],
+            outputRange: [
+              coordinate.latitude,
+              coordinate.latitude - (LATITUDE_DELTA * 0.5 * 0.375),
+            ],
             extrapolate: 'clamp',
           }),
           latitudeDelta: scrollY.interpolate({
@@ -292,11 +293,11 @@ var AnimatedViews = React.createClass({
         }).start();
       }
     }
-  },
+  }
 
-  onRegionChange(region) {
-    //this.state.region.setValue(region);
-  },
+  onRegionChange(/* region */) {
+    // this.state.region.setValue(region);
+  }
 
   render() {
     const {
@@ -317,13 +318,14 @@ var AnimatedViews = React.createClass({
           xMode="snap"
           snapSpacingX={SNAP_WIDTH}
           yBounds={[-1 * screen.height, 0]}
-          xBounds={[-screen.width * (markers.length-1), 0]}
+          xBounds={[-screen.width * (markers.length - 1), 0]}
           panY={panY}
           panX={panX}
           onStartShouldSetPanResponder={this.onStartShouldSetPanResponder}
           onMoveShouldSetPanResponder={this.onMoveShouldSetPanResponder}
         >
           <MapView.Animated
+            provider={this.props.provider}
             style={styles.map}
             region={region}
             onRegionChange={this.onRegionChange}
@@ -374,47 +376,40 @@ var AnimatedViews = React.createClass({
                       { scale },
                     ],
                   }]}
-                >
-                </Animated.View>
+                />
               );
             })}
           </View>
         </PanController>
       </View>
     );
-  },
-});
+  }
+}
 
+AnimatedViews.propTypes = {
+  provider: MapView.ProviderPropType,
+};
 
-
-var styles = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
   },
   itemContainer: {
     backgroundColor: 'transparent',
     flexDirection: 'row',
-    paddingHorizontal: ITEM_SPACING / 2 + ITEM_PREVIEW,
+    paddingHorizontal: (ITEM_SPACING / 2) + ITEM_PREVIEW,
     position: 'absolute',
-    //top: screen.height - ITEM_PREVIEW_HEIGHT - 64,
+    // top: screen.height - ITEM_PREVIEW_HEIGHT - 64,
     paddingTop: screen.height - ITEM_PREVIEW_HEIGHT - 64,
-    //paddingTop: !ANDROID ? 0 : screen.height - ITEM_PREVIEW_HEIGHT - 64,
+    // paddingTop: !ANDROID ? 0 : screen.height - ITEM_PREVIEW_HEIGHT - 64,
   },
   map: {
     backgroundColor: 'transparent',
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    bottom: 0,
-    right: 0,
+    ...StyleSheet.absoluteFillObject,
   },
   item: {
     width: ITEM_WIDTH,
-    height: screen.height + 2 * ITEM_PREVIEW_HEIGHT,
+    height: screen.height + (2 * ITEM_PREVIEW_HEIGHT),
     backgroundColor: 'red',
     marginHorizontal: ITEM_SPACING / 2,
     overflow: 'hidden',
