@@ -1,19 +1,15 @@
-var React = require('react-native');
-var {
+import React from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
   Text,
   Dimensions,
   TouchableOpacity,
-  Image,
-} = React;
+} from 'react-native';
+import MapView from 'react-native-maps';
+import CustomCallout from './CustomCallout';
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
-var CustomCallout = require('./CustomCallout');
-
-var { width, height } = Dimensions.get('window');
+const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -22,9 +18,11 @@ const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 const SPACE = 0.01;
 
-var Callouts = React.createClass({
-  getInitialState() {
-    return {
+class Callouts extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
       region: {
         latitude: LATITUDE,
         longitude: LONGITUDE,
@@ -52,46 +50,49 @@ var Callouts = React.createClass({
         },
       ],
     };
-  },
+  }
 
   show() {
-    this.refs.m1.showCallout();
-  },
+    this.marker1.showCallout();
+  }
 
   hide() {
-    this.refs.m1.hideCallout();
-  },
+    this.marker1.hideCallout();
+  }
 
   render() {
     const { region, markers } = this.state;
     return (
       <View style={styles.container}>
         <MapView
+          provider={this.props.provider}
           style={styles.map}
           initialRegion={region}
         >
           <MapView.Marker
-            ref="m1"
+            ref={ref => { this.marker1 = ref; }}
             coordinate={markers[0].coordinate}
-            title="This is a title"
-            description="This is a description"
+            title="This is a native view"
+            description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation" // eslint-disable-line max-len
           />
-          <MapView.Marker ref="m2" coordinate={markers[1].coordinate}>
-            <MapView.Callout>
+          <MapView.Marker
+            coordinate={markers[1].coordinate}
+          >
+            <MapView.Callout style={styles.plainView}>
+
               <View>
                 <Text>This is a plain view</Text>
               </View>
             </MapView.Callout>
           </MapView.Marker>
           <MapView.Marker
-            ref="m3"
             coordinate={markers[2].coordinate}
             calloutOffset={{ x: -8, y: 28 }}
             calloutAnchor={{ x: 0.5, y: 0.4 }}
           >
-            <MapView.Callout tooltip>
+            <MapView.Callout tooltip style={styles.customView}>
               <CustomCallout>
-                <Text style={{ color: '#fff' }}>This is a custom callout bubble view</Text>
+                <Text>This is a custom callout bubble view</Text>
               </CustomCallout>
             </MapView.Callout>
           </MapView.Marker>
@@ -102,34 +103,37 @@ var Callouts = React.createClass({
           </View>
         </View>
         <View style={styles.buttonContainer}>
-          <TouchableOpacity onPress={this.show} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity onPress={() => this.show()} style={[styles.bubble, styles.button]}>
             <Text>Show</Text>
           </TouchableOpacity>
-          <TouchableOpacity onPress={this.hide} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity onPress={() => this.hide()} style={[styles.bubble, styles.button]}>
             <Text>Hide</Text>
           </TouchableOpacity>
         </View>
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+Callouts.propTypes = {
+  provider: MapView.ProviderPropType,
+};
+
+const styles = StyleSheet.create({
+  customView: {
+    width: 140,
+    height: 100,
+  },
+  plainView: {
+    width: 60,
+  },
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
     justifyContent: 'flex-end',
     alignItems: 'center',
   },
   map: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
   },
   bubble: {
     flex: 1,
