@@ -1,16 +1,15 @@
-var React = require('react-native');
-var {
+import React from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
   Text,
   Dimensions,
   TouchableOpacity,
-} = React;
+} from 'react-native';
 
-var MapView = require('react-native-maps');
+import MapView from 'react-native-maps';
 
-var { width, height } = Dimensions.get('window');
+const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -20,12 +19,14 @@ const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 let id = 0;
 
 function randomColor() {
-  return '#'+Math.floor(Math.random()*16777215).toString(16);
+  return `#${Math.floor(Math.random() * 16777215).toString(16)}`;
 }
 
-var DefaultMarkers = React.createClass({
-  getInitialState() {
-    return {
+class DefaultMarkers extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
       region: {
         latitude: LATITUDE,
         longitude: LONGITUDE,
@@ -34,7 +35,7 @@ var DefaultMarkers = React.createClass({
       },
       markers: [],
     };
-  },
+  }
 
   onMapPress(e) {
     this.setState({
@@ -47,15 +48,16 @@ var DefaultMarkers = React.createClass({
         },
       ],
     });
-  },
+  }
 
   render() {
     return (
       <View style={styles.container}>
         <MapView
+          provider={this.props.provider}
           style={styles.map}
           initialRegion={this.state.region}
-          onPress={this.onMapPress}
+          onPress={(e) => this.onMapPress(e)}
         >
           {this.state.markers.map(marker => (
             <MapView.Marker
@@ -66,31 +68,30 @@ var DefaultMarkers = React.createClass({
           ))}
         </MapView>
         <View style={styles.buttonContainer}>
-          <View style={styles.bubble}>
+          <TouchableOpacity
+            onPress={() => this.setState({ markers: [] })}
+            style={styles.bubble}
+          >
             <Text>Tap to create a marker of random color</Text>
-          </View>
+          </TouchableOpacity>
         </View>
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+DefaultMarkers.propTypes = {
+  provider: MapView.ProviderPropType,
+};
+
+const styles = StyleSheet.create({
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
     justifyContent: 'flex-end',
     alignItems: 'center',
   },
   map: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
   },
   bubble: {
     backgroundColor: 'rgba(255,255,255,0.7)',
