@@ -1,44 +1,52 @@
-var React = require('react-native');
-var {
+import React, { PropTypes } from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
   Text,
   Dimensions,
-  TouchableOpacity,
   ScrollView,
-} = React;
+} from 'react-native';
+// eslint-disable-next-line max-len
+import SyntheticEvent from 'react-native/Libraries/Renderer/src/renderers/shared/stack/event/SyntheticEvent';
+import MapView from 'react-native-maps';
+import PriceMarker from './PriceMarker';
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
-
-var { width, height } = Dimensions.get('window');
+const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
-var id = 0;
+let id = 0;
 
-var Event = React.createClass({
+class Event extends React.Component {
   shouldComponentUpdate(nextProps) {
     return this.props.event.id !== nextProps.event.id;
-  },
+  }
+
   render() {
-    var { event } = this.props;
+    const { event } = this.props;
     return (
       <View style={styles.event}>
         <Text style={styles.eventName}>{event.name}</Text>
         <Text style={styles.eventData}>{JSON.stringify(event.data, null, 2)}</Text>
       </View>
     );
-  },
-});
+  }
+}
 
-var DisplayLatLng = React.createClass({
-  getInitialState() {
-    return {
+Event.propTypes = {
+  event: PropTypes.object,
+};
+
+
+// eslint-disable-next-line react/no-multi-comp
+class EventListener extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
       region: {
         latitude: LATITUDE,
         longitude: LONGITUDE,
@@ -47,33 +55,35 @@ var DisplayLatLng = React.createClass({
       },
       events: [],
     };
-  },
+  }
 
   makeEvent(e, name) {
     return {
       id: id++,
-      name: name,
+      name,
       data: e.nativeEvent ? e.nativeEvent : e,
     };
-  },
+  }
 
   recordEvent(name) {
     return e => {
-      var { events } = this.state;
-      this.setState({
+      if (e instanceof SyntheticEvent && typeof e.persist === 'function') {
+        e.persist();
+      }
+      this.setState(prevState => ({
         events: [
           this.makeEvent(e, name),
-          ...events.slice(0, 10),
+          ...prevState.events.slice(0, 10),
         ],
-      });
+      }));
     };
-  },
+  }
 
   render() {
     return (
       <View style={styles.container}>
         <MapView
-          ref="map"
+          provider={this.props.provider}
           style={styles.map}
           initialRegion={this.state.region}
           onRegionChange={this.recordEvent('Map::onRegionChange')}
@@ -87,6 +97,20 @@ var DisplayLatLng = React.createClass({
           onCalloutPress={this.recordEvent('Map::onCalloutPress')}
         >
           <MapView.Marker
+            coordinate={{
+              latitude: LATITUDE + (LATITUDE_DELTA / 2),
+              longitude: LONGITUDE + (LONGITUDE_DELTA / 2),
+            }}
+          />
+          <MapView.Marker
+            coordinate={{
+              latitude: LATITUDE - (LATITUDE_DELTA / 2),
+              longitude: LONGITUDE - (LONGITUDE_DELTA / 2),
+            }}
+          />
+          <MapView.Marker
+            title="This is a title"
+            description="This is a description"
             coordinate={this.state.region}
             onPress={this.recordEvent('Marker::onPress')}
             onSelect={this.recordEvent('Marker::onSelect')}
@@ -103,6 +127,34 @@ var DisplayLatLng = React.createClass({
               </View>
             </MapView.Callout>
           </MapView.Marker>
+          <MapView.Polygon
+            fillColor={'rgba(255,0,0,0.3)'}
+            onPress={this.recordEvent('Polygon::onPress')}
+            coordinates={[{
+              latitude: LATITUDE + (LATITUDE_DELTA / 5),
+              longitude: LONGITUDE + (LONGITUDE_DELTA / 4),
+            }, {
+              latitude: LATITUDE + (LATITUDE_DELTA / 3),
+              longitude: LONGITUDE + (LONGITUDE_DELTA / 4),
+            }, {
+              latitude: LATITUDE + (LATITUDE_DELTA / 4),
+              longitude: LONGITUDE + (LONGITUDE_DELTA / 2),
+            }]}
+          />
+          <MapView.Polyline
+            strokeColor={'rgba(255,0,0,1)'}
+            onPress={this.recordEvent('Polyline::onPress')}
+            coordinates={[{
+              latitude: LATITUDE + (LATITUDE_DELTA / 5),
+              longitude: LONGITUDE - (LONGITUDE_DELTA / 4),
+            }, {
+              latitude: LATITUDE + (LATITUDE_DELTA / 3),
+              longitude: LONGITUDE - (LONGITUDE_DELTA / 4),
+            }, {
+              latitude: LATITUDE + (LATITUDE_DELTA / 4),
+              longitude: LONGITUDE - (LONGITUDE_DELTA / 2),
+            }]}
+          />
         </MapView>
         <View style={styles.eventList}>
           <ScrollView>
@@ -111,16 +163,19 @@ var DisplayLatLng = React.createClass({
         </View>
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+EventListener.propTypes = {
+  provider: MapView.ProviderPropType,
+};
+
+const styles = StyleSheet.create({
+  callout: {
+    width: 60,
+  },
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
     justifyContent: 'flex-end',
     alignItems: 'center',
   },
@@ -176,4 +231,4 @@ var styles = StyleSheet.create({
   },
 });
 
-module.exports = DisplayLatLng;
+module.exports = EventListener;
