@@ -1,18 +1,15 @@
-var React = require('react-native');
-var {
+import React from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
   Text,
   Dimensions,
-  TouchableOpacity,
-  Image,
-} = React;
+} from 'react-native';
+import MapView from 'react-native-maps';
+import flagBlueImg from './assets/flag-blue.png';
+import flagPinkImg from './assets/flag-pink.png';
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
-
-var { width, height } = Dimensions.get('window');
+const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -21,12 +18,20 @@ const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 const SPACE = 0.01;
 
-var MarkerTypes = React.createClass({
+class MarkerTypes extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      marker1: true,
+      marker2: false,
+    };
+  }
+
   render() {
     return (
       <View style={styles.container}>
         <MapView
-          ref="map"
+          provider={this.props.provider}
           style={styles.map}
           initialRegion={{
             latitude: LATITUDE,
@@ -34,47 +39,63 @@ var MarkerTypes = React.createClass({
             latitudeDelta: LATITUDE_DELTA,
             longitudeDelta: LONGITUDE_DELTA,
           }}
-          >
+        >
           <MapView.Marker
+            onPress={() => this.setState({ marker1: !this.state.marker1 })}
             coordinate={{
               latitude: LATITUDE + SPACE,
               longitude: LONGITUDE + SPACE,
             }}
             centerOffset={{ x: -18, y: -60 }}
             anchor={{ x: 0.69, y: 1 }}
-            image={require('./assets/flag-blue.png')}
-            />
+            image={this.state.marker1 ? flagBlueImg : flagPinkImg}
+          >
+            <Text style={styles.marker}>X</Text>
+          </MapView.Marker>
           <MapView.Marker
+            onPress={() => this.setState({ marker2: !this.state.marker2 })}
             coordinate={{
               latitude: LATITUDE - SPACE,
               longitude: LONGITUDE - SPACE,
             }}
             centerOffset={{ x: -42, y: -60 }}
             anchor={{ x: 0.84, y: 1 }}
-            image={require('./assets/flag-pink.png')}
-            />
+            image={this.state.marker2 ? flagBlueImg : flagPinkImg}
+          />
+          <MapView.Marker
+            onPress={() => this.setState({ marker2: !this.state.marker2 })}
+            coordinate={{
+              latitude: LATITUDE + SPACE,
+              longitude: LONGITUDE - SPACE,
+            }}
+            centerOffset={{ x: -42, y: -60 }}
+            anchor={{ x: 0.84, y: 1 }}
+            opacity={0.6}
+            image={this.state.marker2 ? flagBlueImg : flagPinkImg}
+          />
         </MapView>
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+MarkerTypes.propTypes = {
+  provider: MapView.ProviderPropType,
+};
+
+const styles = StyleSheet.create({
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
     justifyContent: 'flex-end',
     alignItems: 'center',
   },
   map: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
+  },
+  marker: {
+    marginLeft: 46,
+    marginTop: 33,
+    fontWeight: 'bold',
   },
 });
 
