@@ -1,31 +1,39 @@
-var React = require('react-native');
-var {
+import React, { PropTypes } from 'react';
+import {
   StyleSheet,
   View,
   Text,
-} = React;
+} from 'react-native';
 
-var PriceMarker = React.createClass({
-  getDefaultProps() {
-    return {
-      fontSize: 13,
-    };
-  },
+const propTypes = {
+  amount: PropTypes.number.isRequired,
+  fontSize: PropTypes.number,
+};
+
+const defaultProps = {
+  fontSize: 13,
+};
+
+class PriceMarker extends React.Component {
   render() {
+    const { fontSize, amount } = this.props;
     return (
       <View style={styles.container}>
         <View style={styles.bubble}>
           <Text style={styles.dollar}>$</Text>
-          <Text style={[styles.amount, { fontSize: this.props.fontSize }]}>{this.props.amount}</Text>
+          <Text style={[styles.amount, { fontSize }]}>{amount}</Text>
         </View>
         <View style={styles.arrowBorder} />
         <View style={styles.arrow} />
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+PriceMarker.propTypes = propTypes;
+PriceMarker.defaultProps = defaultProps;
+
+const styles = StyleSheet.create({
   container: {
     flexDirection: 'column',
     alignSelf: 'flex-start',
