@@ -1,17 +1,15 @@
-var React = require('react-native');
-var {
+import React from 'react';
+import {
   StyleSheet,
-  PropTypes,
   View,
   Text,
   Dimensions,
   TouchableOpacity,
-} = React;
+} from 'react-native';
+import MapView from 'react-native-maps';
+import PriceMarker from './PriceMarker';
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
-
-var { width, height } = Dimensions.get('window');
+const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -19,9 +17,11 @@ const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 
-var DisplayLatLng = React.createClass({
-  getInitialState() {
-    return {
+class ViewsAsMarkers extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
       region: {
         latitude: LATITUDE,
         longitude: LONGITUDE,
@@ -34,20 +34,21 @@ var DisplayLatLng = React.createClass({
       },
       amount: 99,
     };
-  },
+  }
 
   increment() {
     this.setState({ amount: this.state.amount + 1 });
-  },
+  }
 
   decrement() {
     this.setState({ amount: this.state.amount - 1 });
-  },
+  }
 
   render() {
     return (
       <View style={styles.container}>
         <MapView
+          provider={this.props.provider}
           style={styles.map}
           initialRegion={this.state.region}
         >
@@ -56,34 +57,36 @@ var DisplayLatLng = React.createClass({
           </MapView.Marker>
         </MapView>
         <View style={styles.buttonContainer}>
-          <TouchableOpacity onPress={this.decrement} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity
+            onPress={() => this.decrement()}
+            style={[styles.bubble, styles.button]}
+          >
             <Text style={{ fontSize: 20, fontWeight: 'bold' }}>-</Text>
           </TouchableOpacity>
-          <TouchableOpacity onPress={this.increment} style={[styles.bubble, styles.button]}>
+          <TouchableOpacity
+            onPress={() => this.increment()}
+            style={[styles.bubble, styles.button]}
+          >
             <Text style={{ fontSize: 20, fontWeight: 'bold' }}>+</Text>
           </TouchableOpacity>
         </View>
       </View>
     );
-  },
-});
+  }
+}
 
-var styles = StyleSheet.create({
+ViewsAsMarkers.propTypes = {
+  provider: MapView.ProviderPropType,
+};
+
+const styles = StyleSheet.create({
   container: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
     justifyContent: 'flex-end',
     alignItems: 'center',
   },
   map: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
+    ...StyleSheet.absoluteFillObject,
   },
   bubble: {
     backgroundColor: 'rgba(255,255,255,0.7)',
@@ -108,4 +111,4 @@ var styles = StyleSheet.create({
   },
 });
 
-module.exports = DisplayLatLng;
+module.exports = ViewsAsMarkers;
