@@ -4,7 +4,7 @@
 //
 
 #import <Foundation/Foundation.h>
-#import "RCTView.h"
+#import <React/RCTView.h>
 
 
 @interface AIRMapCallout : RCTView
