@@ -4,7 +4,7 @@
 //
 
 #import "AIRMapCircle.h"
-#import "UIView+React.h"
+#import <React/UIView+React.h>
 
 
 @implementation AIRMapCircle {
