@@ -13,8 +13,8 @@
 #import <MapKit/MapKit.h>
 #import <UIKit/UIKit.h>
 
-#import "RCTConvert+MapKit.h"
-#import "RCTComponent.h"
+#import <React/RCTConvert+MapKit.h>
+#import <React/RCTComponent.h>
 #import "AIRMap.h"
 #import "SMCalloutView.h"
 
@@ -32,6 +32,8 @@
 @property (nonatomic, copy) NSString *subtitle;
 @property (nonatomic, assign) CLLocationCoordinate2D coordinate;
 @property (nonatomic, strong) UIColor *pinColor;
+@property (nonatomic, assign) NSInteger zIndex;
+@property (nonatomic, assign) double opacity;
 
 @property (nonatomic, copy) RCTBubblingEventBlock onPress;
 @property (nonatomic, copy) RCTDirectEventBlock onSelect;
@@ -47,9 +49,10 @@
 - (BOOL)shouldShowCalloutView;
 - (void)showCalloutView;
 - (void)hideCalloutView;
+- (void)addTapGestureRecognizer;
 
 @end
 
 
-@interface EmptyCalloutBackgroundView : SMCalloutBackgroundView
+@interface AIREmptyCalloutBackgroundView : SMCalloutBackgroundView
 @end
