@@ -9,9 +9,9 @@
 
 #import "AIRMapMarkerManager.h"
 
-#import "RCTUIManager.h"
-#import "RCTConvert+CoreLocation.h"
-#import "UIView+React.h"
+#import <React/RCTConvert+CoreLocation.h>
+#import <React/RCTUIManager.h>
+#import <React/UIView+React.h>
 #import "AIRMapMarker.h"
 
 @interface AIRMapMarkerManager () <MKMapViewDelegate>
@@ -25,15 +25,12 @@ RCT_EXPORT_MODULE()
 - (UIView *)view
 {
     AIRMapMarker *marker = [AIRMapMarker new];
-    UITapGestureRecognizer *tapGestureRecognizer = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(_handleTap:)];
-    // setting this to NO allows the parent MapView to continue receiving marker selection events
-    tapGestureRecognizer.cancelsTouchesInView = NO;
-    [marker addGestureRecognizer:tapGestureRecognizer];
+    [marker addTapGestureRecognizer];
     marker.bridge = self.bridge;
     return marker;
 }
 
-//RCT_EXPORT_VIEW_PROPERTY(identifier, NSString)
+RCT_EXPORT_VIEW_PROPERTY(identifier, NSString)
 //RCT_EXPORT_VIEW_PROPERTY(reuseIdentifier, NSString)
 RCT_EXPORT_VIEW_PROPERTY(title, NSString)
 RCT_REMAP_VIEW_PROPERTY(description, subtitle, NSString)
@@ -43,6 +40,8 @@ RCT_EXPORT_VIEW_PROPERTY(calloutOffset, CGPoint)
 RCT_REMAP_VIEW_PROPERTY(image, imageSrc, NSString)
 RCT_EXPORT_VIEW_PROPERTY(pinColor, UIColor)
 RCT_EXPORT_VIEW_PROPERTY(draggable, BOOL)
+RCT_EXPORT_VIEW_PROPERTY(zIndex, NSInteger)
+RCT_EXPORT_VIEW_PROPERTY(opacity, double)
 
 RCT_EXPORT_VIEW_PROPERTY(onPress, RCTBubblingEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onSelect, RCTDirectEventBlock)
@@ -77,42 +76,6 @@ RCT_EXPORT_METHOD(hideCallout:(nonnull NSNumber *)reactTag)
             [marker.map deselectAnnotation:marker animated:NO];
         }
     }];
-}
-
-#pragma mark - Events
-
-- (void)_handleTap:(UITapGestureRecognizer *)recognizer {
-    AIRMapMarker *marker = (AIRMapMarker *)recognizer.view;
-    if (!marker) return;
-
-    if (marker.selected) {
-        CGPoint touchPoint = [recognizer locationInView:marker.map.calloutView];
-        if ([marker.map.calloutView hitTest:touchPoint withEvent:nil]) {
-
-            // the callout got clicked, not the marker
-            id event = @{
-                    @"action": @"callout-press",
-            };
-
-            if (marker.onCalloutPress) marker.onCalloutPress(event);
-            if (marker.calloutView && marker.calloutView.onPress) marker.calloutView.onPress(event);
-            if (marker.map.onCalloutPress) marker.map.onCalloutPress(event);
-            return;
-        }
-    }
-
-    // the actual marker got clicked
-    id event = @{
-            @"action": @"marker-press",
-            @"id": marker.identifier ?: @"unknown",
-            @"coordinate": @{
-                    @"latitude": @(marker.coordinate.latitude),
-                    @"longitude": @(marker.coordinate.longitude)
-            }
-    };
-
-    if (marker.onPress) marker.onPress(event);
-    if (marker.map.onMarkerPress) marker.map.onMarkerPress(event);
 }
 
 @end
