@@ -3,7 +3,7 @@
 // Copyright (c) 2015 Facebook. All rights reserved.
 //
 
-#import "RCTViewManager.h"
+#import <React/RCTViewManager.h>
 
 
 @interface AIRMapPolygonManager : RCTViewManager
